documentのborder変更前

+++ application/controllers/tools.php
@@ -37,19 +37,42 @@ class Tools extends CI_Controller {
 
     public function analyze()
     {
-        $today = date('Ymd',time());
+        //$today = date('Ymd',time());
+        $time = time();
+        $today = date('Y',$time).'/'.date('m',$time).'/'.date('d',$time);
+        list($year,$month,$day) = explode('/',$today);
+
         //zip
         $zip_path = '/usr/local/apache2/htdocs/disclosure/uploads/zip/';
-        $tmp_path = '/usr/local/apache2/htdocs/disclosure/uploads/tmp/'.$today;
-        
-        $move_path = '/usr/local/apache2/htdocs/disclosure/xbrls/'.$today;
+        $tmp_path = '/usr/local/apache2/htdocs/disclosure/uploads/tmp/';
+        $move_path = '/usr/local/apache2/htdocs/disclosure/xbrls/';
+        $tmp_ymd_path = '/usr/local/apache2/htdocs/disclosure/uploads/tmp/'.$year.'/'.$month.'/'.$day;
+        $move_ymd_path = '/usr/local/apache2/htdocs/disclosure/xbrls/'.$year.'/'.$month.'/'.$day;
         $rename_paths = array();
-
         
-        if(is_dir($tmp_path)) $this->_remove_directory($tmp_path);
-        @mkdir($tmp_path);
-        @chown($tmp_path, 'apache');
-        @chmod($tmp_path,0770);
+        if(is_dir($tmp_ymd_path)) $this->_remove_directory($tmp_ymd_path);
+        
+        //dir作成
+        $is_dir_go = FALSE;
+        if( $this->_make_date_directory($tmp_path.$year) && $this->_make_date_directory($tmp_path.$year.'/'.$month) && $this->_make_date_directory($tmp_path.$year.'/'.$month.'/'.$day))$is_dir_go = TRUE;
+        if(!$is_dir_go){
+            log_message('error','tmp dir error'.$tmp_ymd_path);
+            echo 'tmp dir error';
+            die();
+        }
+        
+        
+        
+        
+        if( $this->_make_date_directory($move_path.$year) && $this->_make_date_directory($move_path.$year.'/'.$month) && $this->_make_date_directory($move_path.$year.'/'.$month.'/'.$day))$is_dir_go = TRUE;
+        if(!$is_dir_go){
+            log_message('error','move dir error'.$tmp_ymd_path);
+            echo 'move dir error';
+            die();
+        }
+        //@mkdir($tmp_ymd_path);
+        //@chown($tmp_ymd_path, 'apache');
+        //@chmod($tmp_ymd_path,0770);
 
         $list = scandir($zip_path);
         $is_analyze = FALSE;
@@ -62,7 +85,7 @@ class Tools extends CI_Controller {
                 if($pathinfo['extension'] == 'zip'){
                     $is_analyze = TRUE;
                     $tempFolder = $this->_createTemporaryFolder($today);
-                    $tmp_dir_path = $tmp_path .'/' . $tempFolder;
+                    $tmp_dir_path = $tmp_ymd_path .'/' . $tempFolder;
                     // ZIPファイルをオープン
                     $res = $this->archiver->open($zip_file_path);
                      
@@ -77,8 +100,8 @@ class Tools extends CI_Controller {
                         $this->archiver->close();
                         
                         //renameを記録
-                        //$rename_paths[] = array('zip_path'=>$zip_file_path,'tmp_path'=>$tmp_dir_path . $pathinfo['filename'] . $pathinfo['filename'] , 'move_path'=>$move_path . '/' . $pathinfo['filename']);
-                        $rename_paths[] = array('zip_path'=>$zip_file_path,'tmp_path'=>$tmp_dir_path , 'move_path'=>$move_path . '/' . $tempFolder);
+                        //$rename_paths[] = array('zip_path'=>$zip_file_path,'tmp_path'=>$tmp_dir_path . $pathinfo['filename'] . $pathinfo['filename'] , 'move_path'=>$move_ymd_path . '/' . $pathinfo['filename']);
+                        $rename_paths[] = array('zip_path'=>$zip_file_path,'tmp_path'=>$tmp_dir_path , 'move_path'=>$move_ymd_path . '/' . $tempFolder);
                     }else{
 var_dump('unzip error');
 die();
@@ -87,21 +110,19 @@ die();
             }
         }
         if(!$is_analyze){
-            $this->_remove_directory($tmp_path);
+            $this->_remove_directory($tmp_path.$year);
             die();
         }
-        @mkdir($move_path);
-        @chown($move_path, 'apache');
-        @chmod($move_path,0770);
+        //@mkdir($move_ymd_path);
+        //@chown($move_ymd_path, 'apache');
+        //@chmod($move_ymd_path,0770);
         $insert_data = array();
         $csv_datas = array();
         $base_datas = array();
-        $this->_list_files($tmp_dir_path. '/',$move_path);
-
+        $this->_list_files($tmp_dir_path. '/',$move_ymd_path);
+        
+        $presenters_map = array();
         if(!empty($this->xbrls_informations)){
-            //start transaction manually
-            $this->db->trans_begin();
-
             //先に企業名をDBに挿入 presenter_id が必要なため
             foreach ($this->xbrls_informations as $unzip_dir_name => $xbrls_information){
                 foreach ($xbrls_information as $xbrl_dir_id => $value){
@@ -112,21 +133,24 @@ die();
                         $edinet_code = $value['edinet_code'];
                         $insert_data['presenter'][$edinet_code]['edinet_code'] = $edinet_code;
                         $insert_data['presenter'][$edinet_code]['name'] = $value['presenter_name'];
-                        $security = $this->Security_model->getSecurityByName($value['presenter_name']);
+                        $security = $this->Security_model->getSecurityByName( $this->xbrls_informations[$unzip_dir_name][$xbrl_dir_id]['presenter_name'] );
                         $insert_data['presenter'][$edinet_code]['securities_code'] = empty($security) ? '' : $security->id;
                     }
+                    $presenters_map[$xbrl_dir_id] = $value['presenter_name'];
                 }
             }
         }
-        if(!empty($insert_data['presenter']))$this->db->insert_batch('presenters', $insert_data['presenter']);
+        if(!empty($insert_data['presenter']))$this->db->insert_batch('presenters', $insert_data['presenter']);//myisam
 
         //move 
         foreach ($rename_paths as $rename_path){
             if(is_dir($rename_path['move_path'])) $this->_remove_directory($rename_path['move_path']);
             rename($rename_path['tmp_path'],$rename_path['move_path']);
+            $zip_name = end(explode('/',$rename_path['zip_path']));
+            rename($rename_path['zip_path'],$rename_path['move_path'].'/'.$zip_name);
             //@unlink($rename_path['zip_path']);
         }
-        $this->_remove_directory($tmp_path);
+        $this->_remove_directory($tmp_path.$year);
 
         //実際にファイル解析
         $xbrl_count = 0;
@@ -152,7 +176,7 @@ die();
                     $date = $year.'-'.$month.'-'.$day;
                     //format
                     /*
-                    array(11) {
+                    array(15) {
                       [0]=>
                       string(0) ""
                       [1]=>
@@ -168,20 +192,32 @@ die();
                       [6]=>
                       string(5) "xbrls"
                       [7]=>
-                      string(8) "20140723"
+                      string(4) "2014"
                       [8]=>
-                      string(32) "Xbrl_Search_20140718_102237_test"
+                      string(2) "08"
                       [9]=>
-                      string(8) "S1001O6T"
+                      string(2) "06"
                       [10]=>
-                      string(49) "jpfr-ssr-G08995-000-2014-04-21-01-2014-07-18.xbrl"
+                      string(18) "20140806123925fb71"
+                      [11]=>
+                      string(8) "S1001W63"
+                      [12]=>
+                      string(4) "XBRL"
+                      [13]=>
+                      string(9) "PublicDoc"
+                      [14]=>
+                      string(60) "jpsps070000-asr-001_G07041-000_2014-05-12_01_2014-07-31.xbrl"
                     }
                     */
                     $format_path_ex = explode('/', $xbrl_path);
                     $count = count($format_path_ex);
                     //ファイル命名
-                    $format_path_ex[$count-1] = $date.'_'.$this->xbrls_informations[$unzip_dir_name][$xbrl_dir_id]['edinet_code'];//日付+edinetcode
-                    $format_path = implode('/',array_slice($format_path_ex,6,$count));//xbrlsから
+                    //$format_path_ex[$count-1] = $date.'_'.$this->xbrls_informations[$unzip_dir_name][$xbrl_dir_id]['edinet_code'];//日付+edinetcode
+                    //$format_path = implode('/',array_slice($format_path_ex,6,$count));//xbrlsから
+                    $new_format_arr = array_slice($format_path_ex,6,5);
+                    $new_format_arr[] = $date.'_'.$this->xbrls_informations[$unzip_dir_name][$xbrl_dir_id]['edinet_code'];//日付+edinetcode
+                    $format_path = implode('/',$new_format_arr);//xbrlsからランダム文字列ディレクトリ20140806123925fb71まで
+
                     //最初の1ファイル分だけDBデータ生成
                     if($xbrl_number == 0){
                         $insert_data['document'][$xbrl_dir_id]['edinet_code'] = $this->xbrls_informations[$unzip_dir_name][$xbrl_dir_id]['edinet_code'];
@@ -201,7 +237,8 @@ die();
                             $insert_data['document'][$xbrl_dir_id]['category_id'] = $this->categories[$this->xbrls_informations[$unzip_dir_name][$xbrl_dir_id]['document_name']]->id;
                         }
                         //企業チェック
-                        $presenter = $this->Presenter_model->getPresenterByName($this->xbrls_informations[$unzip_dir_name][$xbrl_dir_id]['presenter_name']);
+                        $presenter_name = $presenters_map[$xbrl_dir_id];//XbrlSearchDlInfo.csvを使う
+                        $presenter = $this->Presenter_model->getPresenterByName($presenter_name);
                         if(!empty($presenter)){
                             $insert_data['document'][$xbrl_dir_id]['presenter_id'] = $presenter->id;
                         }
@@ -220,14 +257,13 @@ die();
                     if($this->is_parse){
                         $xbrl_datas = $this->xbrl_lib->_parseXml($xbrl_path);
                         $csv_datas[$xbrl_count] = $this->xbrl_lib->_makeCsvSqlData($xbrl_datas,$xbrl_path,$insert_data['document_data'][$xbrl_dir_id][$xbrl_number]);
-                        $csv_paths[$xbrl_count] = $xbrl_number > 1 ? $format_path.'_'.$xbrl_number.'.csv' : $format_path.'.csv';
+                        $csv_paths[$xbrl_count] = $xbrl_number > 0 ? $format_path.'_'.$xbrl_number.'.csv' : $format_path.'.csv';
                         $excel_map[$xbrl_count]  = $xbrl_path_loop_number;
-                        $excel_sheet_name[$xbrl_dir_id][$xbrl_path_loop_number] = $format_path_ex[$count-1].'_'.$xbrl_path_loop_number;
+                        $excel_sheet_name[$xbrl_dir_id][$xbrl_path_loop_number] = end($new_format_arr).'_'.$xbrl_path_loop_number;
                         $xbrl_count++;
                     }
                     $xbrl_path_loop_number++;
                 }
-
                 //既に存在していたら除去して個別更新
                 if(!empty($check_xbrl)){
                     $this->db->where('code', $code);
@@ -237,27 +273,33 @@ die();
                     $this->db->delete('document_datas');
 
                     for ($i=0;$i<$xbrl_paths_count;$i++){
+                        $batch_data = array();
                         foreach ($insert_data['document_data'][$xbrl_dir_id][$i] as $value){
                             $value['document_id'] = $check_xbrl->id;
-                            $this->db->insert('document_datas', $value);
+                            $batch_data[] = $value;
                         }
+                        $this->db->insert_batch('document_datas', $batch_data);
                     }
                 }else{
                     //DB追加
                     $insert_data['document'][$xbrl_dir_id]['xbrl_path'] = serialize($xbrl_path_arr);//複数の可能性あり
-                    $document_id = $this->db->insert('documents', $insert_data['document'][$xbrl_dir_id]);
+                    $this->db->insert('documents', $insert_data['document'][$xbrl_dir_id]);
+                    $document_id = $this->db->insert_id();
+                    echo $document_id."\n";
                     for ($i=0;$i<$xbrl_paths_count;$i++){
+                        $batch_data = array();
                         foreach ($insert_data['document_data'][$xbrl_dir_id][$i] as $value){
                             $value['document_id'] = $document_id;
-                            $this->db->insert('document_datas', $value);
+                            $batch_data[] = $value;
                         }
+                        $this->db->insert_batch('document_datas', $batch_data);
                     }
                 }
 
                 
                 //ここでexcel
                 echo $excel_path."\n";
-                //$this->put_excel($excel_path,$csv_datas,$excel_sheet_name[$xbrl_dir_id],$excel_map);
+                $this->put_excel($excel_path,$csv_datas,$excel_sheet_name[$xbrl_dir_id],$excel_map);
             }
             //1documentごとにコミット
             if ($this->db->trans_status() === FALSE) {
@@ -269,13 +311,13 @@ die();
             }
         }
         if($this->is_parse){
-            //$this->put_csv($csv_paths,$csv_datas);
-            //$this->put_csv($base_paths,$base_datas,TRUE);
-            //$this->put_excel($excel_path,$csv_datas,$excel_sheet_name);
+            $this->put_csv($csv_paths,$csv_datas);
         }
+        //最後にディレクトリを削除
+        $this->_remove_directory($rename_path['move_path'],TRUE,array($rename_path['move_path']));
     }
     
-    function _list_files($tmp_dir_path,$move_path){
+    function _list_files($tmp_dir_path,$move_ymd_path){
         $files = array();
         $list = scandir($tmp_dir_path);
         foreach($list as $file){
@@ -286,7 +328,7 @@ die();
                 if($pathinfo['extension'] == 'xbrl'){
                     $dirs = explode('/',$tmp_dir_path);
                     /*
-                    array(14) {
+                    array(16) {
                       [0]=>
                       string(0) ""
                       [1]=>
@@ -304,16 +346,20 @@ die();
                       [7]=>
                       string(3) "tmp"
                       [8]=>
-                      string(8) "20140723"
+                      string(4) "2014"
                       [9]=>
-                      string(18) "20140723155311982f"
+                      string(2) "08"
                       [10]=>
-                      string(8) "S10024IQ"
+                      string(2) "06"
                       [11]=>
-                      string(4) "XBRL"
+                      string(18) "201408061232131ac3"
                       [12]=>
-                      string(9) "PublicDoc"
+                      string(8) "S1001W63"
                       [13]=>
+                      string(4) "XBRL"
+                      [14]=>
+                      string(9) "PublicDoc"
+                      [15]=>
                       string(0) ""
                     }
                     */
@@ -325,7 +371,7 @@ die();
                     $xbrl_add_path = '';
                     foreach ($dirs as $key => $value){
                         if(!empty($value)){
-                            if($key >= 9){
+                            if($key >= 11){
                                 $xbrl_add_path .= '/'.$value;
                             }else{
                                 $base_path .= '/'.$value;
@@ -333,15 +379,15 @@ die();
                         }
 
                     }
-                    $xbrl_path = $move_path.$xbrl_add_path;
-                    $this->xbrl_files[$dirs[9]][$dirs[10]][] = $xbrl_path . '/' . $file;//xbrlが複数ある場合があります
+                    $xbrl_path = $move_ymd_path.$xbrl_add_path;
+                    $this->xbrl_files[$dirs[11]][$dirs[12]][] = $xbrl_path . '/' . $file;//xbrlが複数ある場合があります
                 }elseif($pathinfo['filename'] == 'XbrlSearchDlInfo' && $pathinfo['extension'] == 'csv'){
                     $dirs = explode('/',$tmp_dir_path);
                     //XbrlSearchDlInfo.csv
-                    $this->xbrls_informations[$dirs[9]] = $this->_read_form_edinetinfo_csv($tmp_dir_path . $file);
+                    $this->xbrls_informations[$dirs[11]] = $this->_read_form_edinetinfo_csv($tmp_dir_path . $file);
                 }
             } else if( is_dir($tmp_dir_path . $file) && $file != 'AuditDoc') {
-                $files = array_merge($files, $this->_list_files($tmp_dir_path . $file . '/',$move_path));
+                $files = array_merge($files, $this->_list_files($tmp_dir_path . $file . '/',$move_ymd_path));
             }
         }
         return $files;
@@ -541,7 +587,6 @@ die();
 
     function _createTemporaryFolder($date = '') {
         $folderName = date('YmdHis') . substr(md5(uniqid(mt_rand())), 0, 4);
-        $test = $this->CI->upload_folder->getTemporaryFolder($folderName,$date);
         if ($this->CI->upload_folder->createFolder($this->CI->upload_folder->getTemporaryFolder($folderName,$date))) {
             return $folderName;
         } else {
@@ -550,22 +595,40 @@ die();
         }
     }
 
-    function _remove_directory($dir) {
+    function _remove_directory($dir,$is_directory_only = FALSE,$not_delete_path_arr = array()) {
       if ($handle = opendir("$dir")) {
        while (false !== ($item = readdir($handle))) {
          if ($item != "." && $item != "..") {
            if (is_dir("$dir/$item")) {
-             $this->_remove_directory("$dir/$item");
+             $this->_remove_directory("$dir/$item",FALSE);
            } else {
-             unlink("$dir/$item");
-             //echo " removing $dir/$item<br>\n";
+             if($is_directory_only){
+                 $pathinfo = pathinfo($item);
+                 /*
+                 2014-07-31_G07041.xlsx
+                 XbrlSearchDlInfo.csv
+                 Xbrl_Search_20140805_110324.zip
+                 S1001W63 ← これを削除
+                 */
+                 if($pathinfo['extension'] != 'csv' && $pathinfo['extension'] != 'xlsx' && $pathinfo['extension'] != 'zip'){
+                     unlink("$dir/$item");
+                 }
+             }else{
+                unlink("$dir/$item");
+             }
            }
          }
        }
        closedir($handle);
-       rmdir($dir);
+       if(!in_array($dir,$not_delete_path_arr)) rmdir($dir);
        //echo "removing $dir<br>\n";
       }
+    }
+
+    function _make_date_directory($path) {
+        $bl = FALSE;
+        if( is_dir($path) || ( mkdir($path) && chown($path, 'apache') && chmod($path,0770) ) )$bl = TRUE;
+        return $bl;
     }
 
  public function _parseXml($file="") {
